# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Controle de Estoque API",
        "description": "API para controle de estoque, movimentações e autenticação",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "Leonardo Losso",
            "email": "leonardolosso.sz@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/produtos": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Listar produtos",
                "description": "Retorna uma lista de produtos com filtros opcionais.",
                "operationId": "b44f47e2fa424c65d2db9aba57590885",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Texto para buscar no nome ou descrição do produto",
                        "schema": {
                            "type": "string",
                            "example": "Produto"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filtrar produtos por ID da categoria",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "fornecedor_id",
                        "in": "query",
                        "description": "Filtrar produtos por ID do fornecedor",
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "low_stock",
                        "in": "query",
                        "description": "Filtrar produtos com estoque abaixo do mínimo",
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de produtos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Criar produto",
                "description": "Cria um novo produto.",
                "operationId": "a02ae3401c5ed84c9d6e582ee8e968bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductResource"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": [
                                                {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "O campo name é obrigatório."
                                                    },
                                                    "property": null
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/produtos/{id}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Exibir produto",
                "description": "Retorna os detalhes de um produto específico.",
                "operationId": "cf44ed3aa409d1bd215b1f0054be8952",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto encontrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Atualizar produto",
                "description": "Atualiza os dados de um produto existente.",
                "operationId": "1503bf1a3aff413066d6ac051f559b77",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductResource"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Deletar produto",
                "description": "Remove um produto pelo ID.",
                "operationId": "b117835b428561dfaf68af47eebd24f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Produto deletado com sucesso"
                    },
                    "404": {
                        "description": "Produto não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ProductResource": {
                "required": [
                    "name",
                    "code",
                    "category_id",
                    "fornecedor_id",
                    "cost_price",
                    "sale_price",
                    "min_stock",
                    "stock",
                    "expiry_date"
                ],
                "properties": {
                    "id": {
                        "description": "ID do produto",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nome do produto",
                        "type": "string"
                    },
                    "code": {
                        "description": "Código do produto",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descrição do produto",
                        "type": "string"
                    },
                    "categoryId": {
                        "description": "ID da categoria",
                        "type": "integer",
                        "example": 2
                    },
                    "categoryName": {
                        "description": "Nome da categoria",
                        "type": "string",
                        "example": "Automoveis"
                    },
                    "fornecedorId": {
                        "description": "ID do fornecedor",
                        "type": "integer",
                        "example": 3
                    },
                    "fornecedorName": {
                        "description": "Nome do fornecedor",
                        "type": "string",
                        "example": "Simas Turbo"
                    },
                    "costPrice": {
                        "description": "Preço de custo",
                        "type": "number",
                        "format": "float",
                        "example": 100.5
                    },
                    "salePrice": {
                        "description": "Preço de venda",
                        "type": "number",
                        "format": "float",
                        "example": 150
                    },
                    "expiryDate": {
                        "description": "Data de validade",
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}